# openapi.yaml
openapi: 3.1.0
info:
  title: Meu Advogado 2.0 API
  version: 0.7.0-mobile-signup-ux
servers:
  - url: http://localhost:3333
paths:
  /health:
    get:
      summary: Healthcheck publico da API
      responses:
        "200":
          description: API operacional
  /v1/areas:
    get:
      summary: Lista areas juridicas iniciais
      responses:
        "200":
          description: Areas juridicas retornadas
  /v1/auth/signup-client:
    post:
      summary: Cria usuario cliente
      description: Cria Supabase Auth e profile de dominio com role client. Nao retorna senha, token ou segredo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 120
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
      responses:
        "201":
          description: Usuario cliente criado
        "422":
          description: Payload invalido ou Auth recusou o cadastro
        "503":
          description: Cadastro indisponivel sem credencial server-side
  /v1/auth/change-password:
    post:
      summary: Troca senha do usuario autenticado
      description: Atualiza senha via backend com service role server-side. Nao retorna senha, token ou segredo.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 72
      responses:
        "200":
          description: Senha atualizada e estado de primeiro acesso retornado
        "401":
          description: Token ausente ou invalido
        "422":
          description: Senha invalida ou Auth recusou a atualizacao
  /v1/me:
    get:
      summary: Retorna identidade autenticada segura
      description: Retorna id, email, role e estado de primeiro acesso sem senha ou token.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Usuario autenticado
        "401":
          description: Token ausente ou invalido
        "403":
          description: Perfil sem role autorizada ou usuario bloqueado
  /v1/match:
    post:
      summary: Solicita match por localizacao e areas juridicas
      description: Retorna o advogado aprovado mais proximo e compativel via PostGIS.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng, accuracyM, areaIds]
              properties:
                lat:
                  type: number
                lng:
                  type: number
                accuracyM:
                  type: number
                areaIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Resultado matched ou empty
          content:
            application/json:
              schema:
                type: object
                required: [status, lawyer, algorithmVersion]
                properties:
                  status:
                    type: string
                    enum: [matched, empty]
                  lawyer:
                    anyOf:
                      - type: "null"
                      - type: object
                        required: [id, name, whatsapp, city, state, areaIds]
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          whatsapp:
                            type: string
                          city:
                            type: [string, "null"]
                          state:
                            type: [string, "null"]
                          areaIds:
                            type: array
                            items:
                              type: string
                          avatarUrl:
                            type: [string, "null"]
                            format: uri
                          coverUrl:
                            type: [string, "null"]
                            format: uri
                  distanceKm:
                    type: number
                  algorithmVersion:
                    type: string
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "422":
          description: Payload invalido
  /v1/lawyers/{id}:
    get:
      summary: Retorna perfil profissional publico seguro para cliente autenticado
      description: Retorna somente advogado aprovado. Perfil inexistente ou indisponivel responde 404 seguro.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Allowlist publica do perfil profissional
          content:
            application/json:
              schema:
                type: object
                required: [lawyer]
                properties:
                  lawyer:
                    type: object
                    required: [id, name, oabNumber, oabState, city, state, areaIds, areas, whatsapp, verified]
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      oabNumber:
                        type: string
                      oabState:
                        type: string
                      city:
                        type: [string, "null"]
                      state:
                        type: [string, "null"]
                      areaIds:
                        type: array
                        items:
                          type: string
                      areas:
                        type: array
                        items:
                          type: object
                          required: [id, name]
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                      whatsapp:
                        type: string
                      verified:
                        type: boolean
                        const: true
                      avatarUrl:
                        type: [string, "null"]
                        format: uri
                      coverUrl:
                        type: [string, "null"]
                        format: uri
                      miniBio:
                        type: [string, "null"]
                      fullBio:
                        type: [string, "null"]
                      yearsExperience:
                        type: [integer, "null"]
                      planLabel:
                        type: [string, "null"]
                      emergencyAvailable:
                        type: boolean
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "404":
          description: Perfil inexistente, nao aprovado ou indisponivel
  /v1/admin/lawyers:
    get:
      summary: Lista advogados cadastrados pelo admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista paginada
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
    post:
      summary: Cadastra advogado com CEP obrigatorio e envia convite de acesso
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Advogado criado e convite Auth enviado sem retornar segredo
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "422":
          description: Payload invalido
  /v1/admin/lawyers/{id}/access-invite:
    post:
      summary: Ativa convite de primeiro acesso para advogado legado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Convite enviado e perfil legado revinculado ao auth.users.id
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "404":
          description: Advogado ou perfil nao encontrado
        "409":
          description: Convite ja registrado
        "422":
          description: Auth recusou o convite
  /v1/lawyer/me/dashboard:
    get:
      summary: Retorna dashboard seguro do advogado logado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard com metricas zeradas e beneficios estaticos
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "404":
          description: Dashboard indisponivel
  /v1/prayer-requests:
    post:
      summary: Recebe pedido de oracao do cliente autenticado
      description: Nao ecoa o texto na resposta e aplica validacao de tamanho.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 20
                  maxLength: 500
                anonymous:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Pedido recebido
        "401":
          description: Token ausente ou invalido
        "403":
          description: Role insuficiente
        "422":
          description: Payload invalido
  /v1/admin/lawyers/{id}:
    patch:
      summary: Atualiza advogado por id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Advogado atualizado
        "404":
          description: Advogado nao encontrado
        "422":
          description: Payload invalido
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT